+++ src/WEB-INF/classes/AppserverIo/Apps/Example/Servlets/RandomRuntimeServlet.php
@@ -1,0 +1,64 @@
+<?php
+
+/**
+ * AppserverIo\Apps\Example\Servlets\RandomRuntimeServlet
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @category   Appserver
+ * @package    Apps
+ * @subpackage Example
+ * @author     Tim Wagner <tw@appserver.io>
+ * @copyright  2014 TechDivision GmbH <info@appserver.io>
+ * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link       https://github.com/appserver-io-apps/example
+ * @link       http://www.appserver.io
+ */
+
+namespace AppserverIo\Apps\Example\Servlets;
+
+use AppserverIo\Psr\Servlet\Http\HttpServlet;
+use AppserverIo\Psr\Servlet\Http\HttpServletRequest;
+use AppserverIo\Psr\Servlet\Http\HttpServletResponse;
+
+/**
+ * Demo servlet that produces random processing times.
+ *
+ * @category   Appserver
+ * @package    Apps
+ * @subpackage Example
+ * @author     Tim Wagner <tw@appserver.io>
+ * @copyright  2014 TechDivision GmbH <info@appserver.io>
+ * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link       https://github.com/appserver-io-apps/example
+ * @link       http://www.appserver.io
+ */
+class RandomRuntimeServlet extends HttpServlet
+{
+
+    /**
+     * Handles a HTTP GET request.
+     *
+     * @param \AppserverIo\Psr\Servlet\Http\HttpServletRequest  $servletRequest  The request instance
+     * @param \AppserverIo\Psr\Servlet\Http\HttpServletResponse $servletResponse The response instance
+     *
+     * @return void
+     * @throws \AppserverIo\Psr\Servlet\ServletException Is thrown if the request method is not implemented
+     * @see \AppserverIo\Psr\Servlet\Http\HttpServlet::doGet()
+     */
+    public function doGet(HttpServletRequest $servletRequest, HttpServletResponse $servletResponse)
+    {
+
+        // sleep for a random time between 10 ms - 1000 ms (=== 1 sec)
+        usleep($sleepFor = rand(10000, 1000000));
+
+        // append the processing time
+        $servletResponse->appendBodyStream('Sleept for: ' . $sleepFor);
+    }
+}